pyproject.toml
[project]
name = "Backend"
version = "0.1.0"
description = "The Backend for the Fex Hackathon Project"
authors = [
    { name = "AnthonyLoche", email = "anthonylocheifc@gmail.com" },
    { name = "jonatasperaza", email = "jonatassilvaperaza@gmail.com" },
]
dependencies = [
    "django>=4.2.13",
    "django-rest-framework>=0.1.0",
    "passage-identity>=2.5.0",
    "pydotplus>=2.0.2",
    "django-cors-headers>=4.4.0",
    "openapi-client>=1.1.7",
    "drf-spectacular>=0.27.2",
    "openapi-python-client>=0.21.4",
    "django-extensions>=3.2.3",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
model = "python manage.py graph_models -a -g -o models.png"
